+++ scripts/cleanup-debian.sh
@@ -1,24 +1,26 @@
 #!/usr/bin/env bash
+if [ $1 == 'true' ]
+then
+  # Uninstall Ansible and dependencies.
 
-# Uninstall Ansible and dependencies.
+  pip uninstall ansible
+  apt-get -y remove python-pip python-dev
 
-pip uninstall ansible
-apt-get -y remove python-pip python-dev
+  # Apt cleanup.
 
-# Apt cleanup.
+  apt autoremove
+  apt update
 
-apt autoremove
-apt update
+  # Delete unneeded files.
 
-# Delete unneeded files.
+  rm -f /home/vagrant/*.sh
 
-rm -f /home/vagrant/*.sh
+  # Zero out the rest of the free space using dd, then delete the written file.
 
-# Zero out the rest of the free space using dd, then delete the written file.
+  dd if=/dev/zero of=/EMPTY bs=1M
+  rm -f /EMPTY
 
-dd if=/dev/zero of=/EMPTY bs=1M
-rm -f /EMPTY
+  # Add `sync` so Packer doesn't quit too early, before the large file is deleted.
 
-# Add `sync` so Packer doesn't quit too early, before the large file is deleted.
-
-sync
+  sync
+fi
